requirements.txt
streamlit
pandas
transformers
torch
google-generativeai
google-auth
